refactor: use action class and update tests
